requirements.txt
streamlit
langchain
langchain-core
langchain-community
langchain-ollama
python-docx
faiss-cpu
sentence-transformers
huggingface-hub
regex
